:adhesive_bandage: fix warnings

+++ SuperMarioBros3/Animations.h
@@ -32,9 +32,9 @@ public:
 	CAnimation(int defaultTime = 100) { this->defaultTime = defaultTime; lastFrameTime = -1; currentFrame = -1; }
 	void Add(int spriteId, DWORD time = 0);
 
-	void Render(float x, float y, float direction, int alpha = 255);
+	void Render(float x, float y, int direction, int alpha = 255);
 
-	void RenderFlipY(float x, float y, float direction, int alpha = 255);
+	void RenderFlipY(float x, float y, int direction, int alpha = 255);
 };
 
 typedef CAnimation* LPANIMATION;
